Plot size class distributions

Creates tree size class distribution plots for one or more stands.

size_dist(size_data, stands, bin_size = 10)

Arguments

size_data
Data frame containing size information of trees. Must include the columns: tree_id, stand_id, dbh, and year (i.e. year in which size measurement was taken).

stands
Character vector of names of stands for which size class distributions are desired. To maintain plot clarity it is recommended that 10 stands or fewer are specified.

bin_size
Integer representing desired width of size classes. Default is 10 dbh units.

Value

A single plot containing a panel for each of the specified stands.

Details

The user-provided data frame size_data can contain trees from an unlimited number of stands but it is recommended that no more that 10 stands be specified in the stands argument to maintain resolution in the resulting multi-panel plot. If size_data contains multiple size measurements for a single tree, the tree_id and year columns will be used to select the most recent size measurement for plotting.

Examples

# Single stand
size_dist(size_data = tree, stands = "AB08", bin_size = 5)

#> Error in ggplot(data = size_data, aes(dbh)): could not find function "ggplot"

# Multiple stands
size_dist(size_data = tree, stands = c("AM16", "AO03", "AR07"))

#> Error in ggplot(data = size_data, aes(dbh)): could not find function "ggplot"